src/agent_c/requirements.txt
# services/signatory/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
descope==1.6.0
python-dotenv==1.0.1